requirements.txt
pandas>=2.0.0
python-slugify>=8.0.0
python-dotenv>=1.0.0
zlm>=0.1.0  # ResumeFlow package
openai>=1.12.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
jinja2>=3.1.0
nltk>=3.8.0
scikit-learn>=1.4.0
fpdf>=1.7.0
markdown-pdf>=0.0.1
validators>=0.22.0